requirements.txt
brotlipy==0.7.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.10
cryptography==38.0.4
google-auth==2.3.3
google-auth-oauthlib==0.4.6
idna==3.3
Markdown==3.3.6
nilearn==0.9.1
numpy==1.24.0
oauthlib==3.1.1
Pillow==9.2.0
pip==22.3.1
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.1.0
PySocks==1.7.1
requests==2.27.1
requests-oauthlib==1.3.0
setuptools==65.6.3
tensorboard-data-server==0.6.1
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
typing_extensions==4.4.0
urllib3==1.26.7
Werkzeug==2.0.2
wheel==0.37.1
zipp==3.7.0
